Revert settings page changes.
